@@ -1,6 +1,6 @@
 name: v_chat_input_ui
 description: v chat text input ui and recorder this item is part of v_chat_sdk and it can be used as a standalone package
-version: 2.0.1
+version: 2.2.0
 homepage: https://v-chat-sdk.github.io/vchat-v2-docs/docs/intro
 issue_tracker: https://github.com/v-chat-sdk/v_chat_input_ui/issues
 repository: https://github.com/v-chat-sdk/v_chat_input_ui
@@ -12,27 +12,30 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+  google_maps_flutter: ^2.6.1
+  audio_waveforms: ^1.0.5
   permission_handler: ^11.3.1
   stop_watch_timer: ^3.1.1
   place_picker_v2: ^0.0.2
-  google_maps_flutter: ^2.6.1
-  cross_file: ^0.3.4+1
-  record: 5.1.2
+  #  google_maps_place_picker_mb: ^3.0.1
+  latlong2: ^0.9.1
+  cross_file: ^0.3.4+2
+  #  record_without_android: ^5.1.1
   v_chat_mention_controller: ^1.0.0
-  audio_waveforms: ^1.0.5
-  emoji_picker_flutter: ^2.2.0
-  meta: ^1.12.0
+  emoji_picker_flutter: ^3.1.0
+  meta: ^1.15.0
+  record: 5.1.2
   intl: ^0.19.0
-  v_platform: ^2.0.0
-  cached_network_image: ^3.3.1
-  adaptive_dialog: ^2.1.0
-  phosphor_flutter: ^1.4.0
-  file_picker: ^8.0.3
-  image_cropper: ^5.0.1
+  v_platform: ^2.1.4
+  cached_network_image: ^3.4.1
+  adaptive_dialog: ^2.2.1+2
+  path: ^1.9.0
+  phosphor_flutter: ^2.1.0
+  file_picker: ^8.1.3
+  image_cropper: ^8.1.0
   wechat_camera_picker: ^4.3.4
-  uuid: ^4.4.0
-  path_provider: ^2.1.3
-  path:
+  uuid: ^4.5.1
+  path_provider: ^2.1.5
 dev_dependencies:
   flutter_lints: ^4.0.0
   lints: ^4.0.0
